# openapi.yaml
openapi: 3.0.3
info:
  title: QR Reward Platform API
  version: '1.0'
  description: |
    Promoters fund QR campaigns; end users scan and sign up with a publisher; coins move
    automatically, exactly once. See `SYSTEM_FLOW.md` for the end-to-end narrative behind
    these endpoints and the invariants each one enforces.

    ### Auth schemes
    - **sessionAuth** — `Authorization: Bearer <jwt>` from `/v1/auth/login` or `/v1/auth/signup`.
      12h expiry. Used by promoter/publisher portal routes and the admin console.
    - **apiKeyAuth** — `Authorization: Bearer pk_<hex>` issued once at publisher signup (or via
      rotate). Used only by the publisher's own server, never a browser.
    - Unauthenticated routes: `/healthz`, `GET /r/:code`, `GET /v1/qr-codes/:id/image`,
      `POST /v1/auth/signup`, `POST /v1/auth/login`.
servers:
  - url: http://localhost:4000
    description: Local dev
  - url: '{baseUrl}'
    description: Deployed
    variables:
      baseUrl:
        default: https://api.example.com
security: []
...
paths:
...
  /v1/partnerships:
    post:
      tags: [Portal]
      summary: Request a partnership with a publisher
      description: Promoter only. Sets the commercial terms for every campaign under it.
      security: [{ sessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [publisher_org_id]
              properties:
                publisher_org_id: { type: string, format: uuid }
                coin_rate: { type: integer, minimum: 1, maximum: 100000, default: 50 }
                guest_rate: { type: integer, minimum: 0, description: 'Must be <= coin_rate', default: 10 }
                grace_days: { type: integer, minimum: 0, maximum: 365, default: 7 }
      responses:
        '201':
          description: Partnership created with status `pending`
          content: { application/json: { schema: { $ref: '#/components/schemas/Partnership' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
    get:
      tags: [Portal]
      summary: List partnerships involving my org
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Partnership' } }
...
  /v1/campaigns:
    post:
      tags: [Portal]
      summary: Create a campaign under an active partnership
      description: Promoter only.
      security: [{ sessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [partnership_id, name]
              properties:
                partnership_id: { type: string, format: uuid }
                name: { type: string }
      responses:
        '201':
          description: Campaign created with status `active` and zero budget
          content: { application/json: { schema: { $ref: '#/components/schemas/Campaign' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
    get:
      tags: [Portal]
      summary: List campaigns visible to my org
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Campaign' } }
...
  /v1/campaigns/{id}:
    patch:
      tags: [Portal]
      summary: Pause, resume, or end a campaign
      description: Promoter only, owner of the campaign.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties: { status: { type: string, enum: [active, paused, ended] } }
      responses:
        '200':
          description: Updated campaign
          content: { application/json: { schema: { $ref: '#/components/schemas/Campaign' } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/campaigns/{id}/qr-codes:
    post:
      tags: [Portal]
      summary: Create a QR code for a campaign
      description: Promoter only, owner of the campaign. Time-bound by default (30 days).
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                style: { $ref: '#/components/schemas/QrStyle' }
                expires_in_days:
                  type: integer
                  minimum: 0
                  maximum: 3650
                  default: 30
                  description: '0 = never expires'
                max_uses:
                  type: integer
                  minimum: 1
                  nullable: true
                  description: 'Omit/null = unlimited'
      responses:
        '201':
          description: QR code created
          content: { application/json: { schema: { $ref: '#/components/schemas/QrCode' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
    get:
      tags: [Portal]
      summary: List QR codes for a campaign
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/QrCode' } }
...
  /v1/qr-codes/{id}:
    patch:
      tags: [Portal]
      summary: Restyle a QR code
      description: Promoter only, owner of the code. Cosmetic only — does not touch expiry or use limits.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [style]
              properties: { style: { $ref: '#/components/schemas/QrStyle' } }
      responses:
        '200':
          description: Updated
          content: { application/json: { schema: { $ref: '#/components/schemas/QrCode' } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/redemptions:
    get:
      tags: [Portal]
      summary: My org's recent redemptions (portal view)
      description: Latest 100, either side of the partnership.
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Redemption' } }

  # ---------------- Partner API ----------------
...
  /v1/admin/overview:
    get:
      tags: [Admin]
      summary: Platform-wide stats and ledger integrity check
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                description: 17 aggregates plus `ledger_balanced` (true iff every ledger ref sums to zero) and `conversion_rate`
  /v1/admin/orgs:
    get:
      tags: [Admin]
      summary: List every org (search by name/email)
      security: [{ sessionAuth: [] }]
      parameters:
        - name: q
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/AdminOrg' } }
  /v1/admin/orgs/{id}:
    patch:
      tags: [Admin]
      summary: Suspend/reinstate an org, or edit its name/landing URL
      description: Audited to `GET /v1/admin/audit-log` as `org.patch`.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                suspended: { type: boolean }
                name: { type: string }
                landing_url: { type: string }
                reason: { type: string, description: Audit trail only }
      responses:
        '200':
          description: Updated
          content: { application/json: { schema: { $ref: '#/components/schemas/Org' } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/admin/orgs/{id}/rotate-key:
    post:
      tags: [Admin]
      summary: Rotate a publisher's API key on their behalf
      description: Support path — key leaked or lost. Audited as `org.rotate_key`.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201':
          description: New key, shown once
          content:
            application/json:
              schema: { type: object, properties: { api_key: { type: string } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/admin/orgs/{id}/offboard:
    post:
      tags: [Admin]
      summary: Suspend an org, revoke its API key, and end every campaign it touches — in one action
      description: Audited as `org.offboard`.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { reason: { type: string } }
      responses:
        '201':
          description: Offboarded
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  name: { type: string }
                  type: { type: string }
                  campaigns_ended: { type: integer }
                  api_key_revoked: { type: boolean }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/admin/partnerships:
    get:
      tags: [Admin]
      summary: List every partnership across all tenants
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Partnership' } }
  /v1/admin/partnerships/{id}:
    patch:
      tags: [Admin]
      summary: Override a partnership's rates or status
      description: Audited as `partnership.patch`. `guest_rate` may never exceed `coin_rate`.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                coin_rate: { type: integer, minimum: 1, maximum: 100000 }
                guest_rate: { type: integer, minimum: 0 }
                grace_days: { type: integer, minimum: 0, maximum: 365 }
                status: { type: string, enum: [pending, active] }
      responses:
        '200':
          description: Updated
          content: { application/json: { schema: { $ref: '#/components/schemas/Partnership' } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/admin/campaigns:
    get:
      tags: [Admin]
      summary: List every campaign across all tenants
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Campaign' } }
  /v1/admin/campaigns/{id}:
    patch:
      tags: [Admin]
      summary: Override a campaign's status
      description: Audited as `campaign.patch`.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, paused, ended] }
                reason: { type: string }
      responses:
        '200':
          description: Updated
          content: { application/json: { schema: { $ref: '#/components/schemas/Campaign' } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/admin/campaigns/{id}/kill:
    post:
      tags: [Admin]
      summary: End a campaign and void every code it ever issued, in one action
      description: Audited as `campaign.kill`.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { reason: { type: string } }
      responses:
        '201':
          description: Killed
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  name: { type: string }
                  status: { type: string, example: ended }
                  codes_voided: { type: integer }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/admin/campaigns/{id}/adjust:
    post:
      tags: [Admin]
      summary: Manually adjust a campaign's budget (goodwill credit or clawback)
      description: |
        Goes through the same double-entry ledger as normal funding. Can never push a balance
        below zero. Audited as `campaign.adjust`.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [coins]
              properties:
                coins: { type: integer, description: 'Non-zero, within ±10,000,000. Negative = clawback.' }
                reason: { type: string }
      responses:
        '201':
          description: New budget
          content:
            application/json:
              schema:
                type: object
                properties: { budget: { type: integer }, reason: { type: string, nullable: true } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/admin/qr-codes/{id}:
    patch:
      tags: [Admin]
      summary: Override a QR code's expiry, use limit, or void state
      description: |
        The only way to loosen a code (extend expiry, raise/remove use limit, un-void) —
        promoters may only tighten their own codes. Audited as `qr_code.override`.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                expires_at: { type: string, format: date-time, nullable: true, description: 'null clears the expiry' }
                max_uses: { type: integer, nullable: true, description: 'null clears the limit' }
                voided: { type: boolean }
                reason: { type: string }
      responses:
        '200':
          description: Updated
          content: { application/json: { schema: { $ref: '#/components/schemas/QrCode' } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/admin/qr-codes:
    get:
      tags: [Admin]
      summary: List every QR code across all tenants
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/QrCode' } }
...
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 12h session token from `/v1/auth/signup` or `/v1/auth/login`
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: 'pk_<hex>'
      description: Publisher's Partner API key — server-to-server only
  responses:
    BadRequest:
      description: Validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
...
    Partnership:
      type: object
      properties:
        id: { type: string, format: uuid }
        promoter_org_id: { type: string, format: uuid }
        publisher_org_id: { type: string, format: uuid }
        promoter_name: { type: string }
        publisher_name: { type: string }
        coin_rate: { type: integer }
        guest_rate: { type: integer }
        grace_days: { type: integer }
        status: { type: string, enum: [pending, active] }
        created_at: { type: string, format: date-time }
...